Change player position to float
Former-commit-id: 63d6460b93f22e56f31131030be6c2a2982bb1eb

+++ server/packet.go
@@ -27,5 +27,5 @@ type Join struct {
 
 type Input struct {
     Id string
-    Move int
+    Move float64
 }
